add parameter 'preset' to the custom derive IntoResponse
In the previous method using 'with = "..", that takes the path to a
free function with the same signature as `IntoResponse::into_response`,
it was impossible to properly set the associated type since the
information about the specified function is unknown.

This patch introduces a new trait `Preset<T>` that contains information
necessary for code generation, and a parameter 'preset = ".."' to
the custom derive `IntoResponse` to specify the code generation using
this trait.